+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# 2.0.0 / 2017-02-09
+
+* Format synthetics.json file to make it human readable
+* Add Developer documentation
+* Add Contribution guide
+* Add ability to update alerting
+* Add command to get available locations
+* Add ability to update synthetics configuration
+* Add support for non-SCRIPTED_BROWSER synthetics
+* Better handling of New Relic errors
+
 # 1.0.1 / 2017-01-04
 
 * Updated README